BugID(s): VOIP-64473
3.2.41 Jun 01 2012
- msvc gtest fixes
- refactoring of screen, call, sipp in prep for win32 port



git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@31800 85c8ee13-dc6b-4ec3-b49e-93a6737f2991

+++ win32_compatibility.hpp
@@ -29,7 +29,7 @@ struct timezone {
 int gettimeofday(struct timeval *tv, struct timezone *tz);
 
 int strcasecmp(const char *str1, const char *str2);
-int strncasecmp(const char *str1, const char *str2, int n);
+int strncasecmp(const char *str1, const char *str2, size_t n);
 const char *strcasestr(const char *s, const char *pattern);
 char *strcasestr(char *s, const char *pattern);
